Removing nucleotide contig fragment context.
