+++ adam-core/src/test/scala/org/bdgenomics/adam/rdd/pileup/PileupRDDFunctionsSuite.scala
@@ -18,7 +18,7 @@
 package org.bdgenomics.adam.rdd.pileup
 
 import org.apache.spark.rdd.RDD
-import org.bdgenomics.adam.rdd.pileup.PileupContext._
+import org.bdgenomics.adam.rdd.ADAMContext._
 import org.bdgenomics.adam.util.ADAMFunSuite
 import org.bdgenomics.formats.avro._
